fix: assembler offset memory bug and label fill bug fixed

+++ assembler/src/parser/directive.go
@@ -3,8 +3,8 @@ package parser
 type Directive struct {
 	raw      string
 	code     string
-	position uint64
-	offset   uint64
+	position uint16
+	offset   uint16
 	single   bool
 	inm      []uint16
 }
